+++ myclub/src/main/java/com/lmbibbo/service/PlayerService.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface PlayerService {
     Player create(Player player);
     Collection<Player> list(int limit);
-    Optional<Player> get(Integer id);
+    Optional<Player> get(Long id);
     Player update(Player player);
-    Boolean delete(Integer id);
+    Boolean delete(Long id);
 }
